src/app/api/requirements.txt
agentmail==0.0.60
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
bech32==1.2.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cosmpy==0.11.1
distlib==0.4.0
distro==1.9.0
ecdsa==0.19.1
filelock==3.19.1
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.39.1
googleapis-common-protos==1.70.0
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.6.4
openai==1.109.1
platformdirs==4.4.0
propcache==0.3.2
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uagents==0.22.9
uagents-core==0.3.9
urllib3==2.5.0
uvicorn==0.37.0
virtualenv==20.34.0
websockets==15.0.1
yarl==1.20.1
